Fix/ff state handling (#430)
Problem: FFs' internal state can be part of a complex equation computing the next state. This is not accounted for in current HAL architecture.

Fix: Add state component to FFs and Latches that stores identifiers for the internal states. Take these identifiers into account when parsing any Boolean function corresponding to a FF or Latch. Update gate libraries to feature the respective information.

ToDo: Update HAL simulator to take state names into account when evaluating Boolean functions. This part is paused for the moment to allow for bigger changes to the simulator plugin that are currently underway.